it trains!

+++ src/masks/utils.py
@@ -8,5 +8,5 @@ def apply_masks(x, masks):
     all_x = []
     for m in masks:
         mask_keep = m.unsqueeze(-1).repeat(1, 1, x.size(-1))
-        all_x += [torch.gather(x, dim=1, index=mask_keep)]
+        all_x += [torch.gather(x, dim=1, index=mask_keep.long())]
     return torch.cat(all_x, dim=0)
